MAINT: Moved all lsq-related files to _lsq subpackage
