+++ scipy/optimize/__init__.py
@@ -235,7 +235,7 @@ from .nnls import nnls
 from ._basinhopping import basinhopping
 from ._linprog import linprog, linprog_verbose_callback
 from ._differentialevolution import differential_evolution
-from .least_squares import least_squares
+from ._lsq import least_squares
 
 __all__ = [s for s in dir() if not s.startswith('_')]
 from numpy.testing import Tester
