+++ scipy/optimize/_lsq/__init__.py
@@ -1,0 +1,7 @@
+"""Module contains last-squares algorithms."""
+
+from __future__ import division, print_function, absolute_import
+
+from .least_squares import least_squares
+
+__all__ = ['least_squares']
