+++ scipy/optimize/setup.py
@@ -71,6 +71,8 @@ def configuration(parent_package='',top_path=None):
 
     config.add_extension('_group_columns', sources=['_group_columns.c'],)
 
+    config.add_subpackage('_lsq')
+
     config.add_data_dir('tests')
     return config
